Fix compilation errors in disabled functions and type conversion
COMPILATION FIXES:
- PatternEngine.cpp:186 - Added return statement to disabled analyzeBalance() function
- PatternEngine.cpp:228 - Added return statement to disabled calculateCenterOfGravityAngle() function
- UPIParser.cpp:1310 - Fixed implicit conversion warning with static_cast<int>()

TECHNICAL DETAILS:
- analyzeBalance() and calculateCenterOfGravityAngle() have commented-out implementations
- Added proper return statements to maintain function contracts
- Fixed size_type to int conversion in progressive pattern cleanup
- All functions now compile without errors while preserving disabled state

Code compiles cleanly with all functionality preserved.

+++ Plugin/Source/UPIParser.cpp
@@ -1307,7 +1307,7 @@ static void cleanupProgressiveStates()
     std::sort(accessCounts.begin(), accessCounts.end());
     
     // Remove the least used patterns (keep only the most recent half)
-    int toRemove = progressivePatterns.size() - (MAX_PROGRESSIVE_STATES / 2);
+    int toRemove = static_cast<int>(progressivePatterns.size()) - (MAX_PROGRESSIVE_STATES / 2);
     for (int i = 0; i < toRemove && i < accessCounts.size(); ++i)
     {
         const juce::String& keyToRemove = accessCounts[i].second;
